Implement carousel page for now.
I realised there are technical challenges I might not be able to
overcome in the amount of time I foresaw for the project. Since all
depends on the amount of technical problems faced it is more important
right now to get an overview of these problems by prototyping instead of
first getting the layout skeleton of each page even though this might make it
more laborious to do it this way around.