Restructure models of challenges
An api will be implemented to make clean ajax requests. The
problem is that with the previous structure a user could not
get his role instance straight out of a challenge model via fields.

On the server side a user could've just used the custom `get_user_role`
method of the current challenge to get its role. But I haven't found a
way to do this via django-rest-framework yet (if even possible).

Another solution could have been to work with ContentTypes. But
the structure in this commit seems the least complex solution
to the problem (import this!).