Comments model at the crossroads
Implement two comment models (HeadComment and TailComment) or one
(Comment)?

The downside of having only one Comment model is that I will have to
implement a model method that returns the Comment instances that belong
to the same head comment. Otherwise I won't be able to render tail
comments under the head comment it belongs to.
It's a cleaner archetechture to have two seperate models
because I wouldn't need this model method then. Also,
I'm using 2 different forms anyway, so it wouldn't make
that much difference in my views.

On the other hand having 2 comment models will result
in tail_comment_set and head_comment_set methods for
Jurors as well as Clenchers, while Clenchers can't
make head_comments. So this would misrepresent the
data structure.

An argument that could be made is why I don't just split
the role model up in Clencher and Juror models. Well,
I need to be able to get the role of the current user
 for the current challenge in the template via js.
But the problem is as that as far as I know you can't define
a serializer field that uses a custom model method and
allows you to pass any arguments(in
this case the user or user id) to craft the api call you'd need
to make.